Use XMLSize_t instead of size_t except where interfacing with other code that uses size_t (e.g., operator new in the C++ runtime).
git-svn-id: https://svn.apache.org/repos/asf/xerces/c/trunk@673975 13f79535-47bb-0310-9956-ffa450edef68

+++ src/xercesc/framework/StdOutFormatTarget.cpp
@@ -39,12 +39,7 @@ void StdOutFormatTarget::writeChars(const XMLByte* const  toWrite
                                   , const XMLSize_t       count
                                   , XMLFormatter* const)
 {
-    // Surprisingly, Solaris was the only platform on which
-    // required the char* cast to print out the string correctly.
-    // Without the cast, it was printing the pointer value in hex.
-    // Quite annoying, considering every other platform printed
-    // the string with the explicit cast to char* below.
-    fwrite(toWrite, sizeof(XMLByte), (size_t)count, stdout);
+    fwrite(toWrite, sizeof(XMLByte), count, stdout);
     fflush(stdout);
 }
